# short circuit anynull if eltype !<: Nullable