--- a/src/primitives.jl
+++ b/src/primitives.jl
@@ -217,7 +217,7 @@ function anynull{T}(X::AbstractArray{T})                       # -> Bool
     if !u && !(Nullable <: T) && !(T <: Nullable)
         return false
     elseif u
-        if VERSION < v"0.6.0-" && !any(S -> S <: Nullable || Nullable <: S, T.types)
+        @static if VERSION < v"0.6.0-" && !any(S -> S <: Nullable || Nullable <: S, T.types)
             return false
         elseif VERSION >= v"0.6.0-" && !any(S -> S <: Nullable || Nullable <: S, Base.uniontypes(T))
             return false