example
```julia
using NullableArrays
using BenchmarkTools
X = rand(Int(1e6));
Y = rand(Int(1e4), Int(1e4));
@benchmark anynull(X)
@benchmark anynull(Y)
```

before
```julia
julia> @benchmark anynull(X)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     270.964 μs (0.00% GC)
  median time:      396.844 μs (0.00% GC)
  mean time:        373.259 μs (0.00% GC)
  maximum time:     917.353 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark anynull(Y)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     27.749 ms (0.00% GC)
  median time:      29.160 ms (0.00% GC)
  mean time:        29.763 ms (0.00% GC)
  maximum time:     36.749 ms (0.00% GC)
  --------------
  samples:          168
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

after
```julia
julia> @benchmark anynull(X)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     18.482 ns (0.00% GC)
  median time:      19.022 ns (0.00% GC)
  mean time:        20.409 ns (0.00% GC)
  maximum time:     87.697 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     997
  time tolerance:   5.00%
  memory tolerance: 1.00%

julia> @benchmark anynull(Y)
BenchmarkTools.Trial:
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     17.400 ns (0.00% GC)
  median time:      17.926 ns (0.00% GC)
  mean time:        19.375 ns (0.00% GC)
  maximum time:     82.290 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     998
  time tolerance:   5.00%
  memory tolerance: 1.00%
```