@cjprybol 2017-03-23
I'm not sure what the coverage drop is for in regards to coveralls, but I could open another PR to try and extend test coverage?