@ararslan 2017-03-23
Weird that Codecov went up negligibly and Coveralls fell significantly... not sure how to interpret that. I guess I'm inclined to call it a fluke, but I'll leave that up to @nalimilan.