@cjprybol 2017-03-24
All tests but coveralls are happy again. If @nalimilan is ok with these changes when the sun comes up on his side of the world, then we can merge and I'll followup with another PR to make the same changes for dropnull & dropnull! and add appropriate tests. Thanks for the help and reviews to both of you!